Dockerfile/nginx.conf: switch to openresty:latest for a newer nginx release
In order to support HTTP/2 nginx has to be at least in version 1.9.5.
That version is not included in Debian repositories. On the other hand,
the official image does not bundle Lua modules that we require. Hence,
switching to openresty which is an updated nginx with a bundle of Lua
modules.

Since openresty is installed in /usr/local/openresty, some paths had to
be updated in our config file.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>