nginx: userauth: disable client_max_body_size check for /userauth
Intergrations API requests are routed through internal endpoint /userauth.

When a POST to /deployments/images hits this endpoint, the subrequset to
/useauth fails due to client_max_body_size check. This is hard to explain, as
the endpoint is configured to *drop* the body of original request. As a
workaround, allow body of any size, but keep the policy of removing it from a
forwarded auth request.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>