Use temporary redirects for localhost.
Use temporary redirects if the host name is localhost, otherwise any
other service the user launches at that address will not be reachable
because the browser auto-redirects to the gateway address. It's quite
tricky to reset this once it's happened.

Discovered when trying to view documentation locally inside a Docker
host; the browser was trying to access the gateway which was long
since torn down.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>