nginx: add Access-Control-Allow-Origin to every 401 status response
Responses carrying 4xx codes will be blocked by CORS functionality of the
browser or requests are made to a different origin. The response is expected to
carry Access-Control-Allow-Origin header that indicates if Origin used in the
original request is allowed to access particular resource. At the javascript
side, when blocked, xhr will not contain information about request status, in
particular status code will be 0.

Use `Headers More` module to set Access-Control-Allow-Origin on every 401
response. This should not cause problems for requests that are not CORS, and
will enable CORS requests to work. The setting is applied globally in `server`
block.

Fixes: [Mender #962]

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>