nginx: adjust proxy settings for deviceauth URL changes
Device Authentication service URLs are now aligned wit the external API URLs.
Meaning that for management API URI we can just prefix match anything that comes
in and pass it directly to the services. For device URIs, use full URIs in
`location` blocks and similarly for device token verification endpoint.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>