nginx.conf: add a default server to prevent host header injection attacks
background: http://nginx.org/en/docs/http/request_processing.html

nginx will first match a request port and try to find the correct server
block by server_name (i.e. Host header). if Host can't be matched (empty, unknown),
nginx will handle the request with the default server block (first in config for a given port)

as a result, all attempts to access 80 or 443 without a known Host header will now
end up in the default server with http 444, which signals to nginx to terminate the
connection (nonstandard code).

for the sake of completeness, the default server still has ssl keys specified.
without them, the connection would still be terminated, but this time due to an internal error,
and not our choice, so we don't want to rely on that.

Issues: MEN-1262

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>