nginx.conf: add a configurable Host whitelist
now that we have a catchall default server, denying requests with unknown Hosts,
we have to be specific about which Hosts we actually allow. this will depend on the environment,
so hide it behind a template var/env var.

the entrypoint picks up the ALLOWED_HOSTS env var and adds them via template
substitution in nginx.conf before runnig nginx.

this way the whitelist can be customized via environment in compose files depending
on a particular deployment.

Issues: MEN-1262

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>