nginx.conf: introduce 'resolver' + variable urls in proxy pass
this is the only real way of getting dns reloading from free nginx.
anything else will cause nginx to resolve dns names once only - at startup or reload.

this is stated in the following doc (disregard the title, it has info on the
free version as well):
https://www.nginx.com/blog/dns-service-discovery-nginx-plus

this is a poc just for deviceauth, should be enough for testing the waters.
(plus it's easy to debug with the qemu client constantly pinging the service in 'integration').

Issues: MEN-1227

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>