reload-when-hosts-changed: a dns/ip monitor for periodic dns cache reloading
since the 'resolver' approach didn't work, this is a workaround.

a standalone 'reload-when-hosts-changed' script is introduced which periodically gets
ip addresses of known services. if address changes are detected compared to a
previous run, '-s reload' is sent to openresty.

by default, services included in the on-prem distribution are monitored.
the list of service names can be overriden via the $DNS_NAMES env var.

Issues: MEN-1227

Changelog: gateway dns cache reloading for improved recovery from service restarts

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>