nginx.conf: disable Strict Transport Security by default
new browsers have strict checks and won't allow the self-signed cert
when hsts is on (no way to override this). this means no access in
demo environment.

set the max-age to 0 to reset/disable hsts, and allow users to
override ssl warnings in demo environments.

note that production template config still has hsts - certs must be
valid in production so that's not an issue.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>